Make editability remanent using viewmodel
This benefits principe of least astonishment, as the viewmodel values
are persisted across hot code pushes but the editable settings, up to
now, weren't. As read-only widget templates lack a viewmodel to begin
with, we had this weird bug whence after a code push, switching the
widget back into editable state would "resurrect" the data it had
prior to the push.